refactor(suite-native): extract connect and unlock screen content to a new component
