+++ suite-native/module-authorize-device/src/screens/connect/ConnectAndUnlockDeviceScreen.tsx
@@ -1,5 +1,4 @@
 import { useCallback, useEffect } from 'react';
-import { Dimensions } from 'react-native';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { useIsFocused, useNavigation } from '@react-navigation/native';
@@ -9,34 +8,16 @@ import {
     selectIsDeviceAuthorized,
     selectIsDeviceConnected,
 } from '@suite-common/wallet-core';
-import { Text, VStack } from '@suite-native/atoms';
-import { ConnectDeviceAnimation } from '@suite-native/device';
+import { ConnectAndUnlockDeviceScreenContent } from '@suite-native/device';
 import { requestPrioritizedDeviceAccess } from '@suite-native/device-mutex';
-import { Translation } from '@suite-native/intl';
 import {
     AuthorizeDeviceStackParamList,
     AuthorizeDeviceStackRoutes,
     Screen,
     StackProps,
 } from '@suite-native/navigation';
-import { prepareNativeStyle, useNativeStyles } from '@trezor/styles';
 
 import { ConnectDeviceScreenHeader } from '../../components/connect/ConnectDeviceScreenHeader';
-
-const ANIMATION_HEIGHT = Dimensions.get('screen').height * 0.6;
-
-const screenContentStyle = prepareNativeStyle(utils => ({
-    flex: 1,
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingTop: utils.spacings.sp40,
-}));
-
-const animationStyle = prepareNativeStyle(() => ({
-    // Both height and width has to be set https://github.com/lottie-react-native/lottie-react-native/blob/master/MIGRATION-5-TO-6.md#updating-the-style-props
-    height: ANIMATION_HEIGHT,
-    width: '100%',
-}));
 
 export const ConnectAndUnlockDeviceScreen = ({
     route: { params },
@@ -44,7 +25,6 @@ export const ConnectAndUnlockDeviceScreen = ({
     AuthorizeDeviceStackParamList,
     AuthorizeDeviceStackRoutes.ConnectAndUnlockDevice
 >) => {
-    const { applyStyle } = useNativeStyles();
     const dispatch = useDispatch();
 
     const isDeviceAuthorized = useSelector(selectIsDeviceAuthorized);
@@ -83,12 +63,7 @@ export const ConnectAndUnlockDeviceScreen = ({
             hasBottomInset={false}
             isScrollable={false}
         >
-            <VStack style={applyStyle(screenContentStyle)}>
-                <Text variant="titleMedium" textAlign="center">
-                    <Translation id="moduleConnectDevice.connectAndUnlockScreen.title" />
-                </Text>
-                <ConnectDeviceAnimation style={applyStyle(animationStyle)} />
-            </VStack>
+            <ConnectAndUnlockDeviceScreenContent />
         </Screen>
     );
 };
